requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
pyvis>=0.3.2
networkx>=3.0
